Moved and documented last Layer specific functions from QgsSpatiaLiteProvider to SpatialiteDbLayer/QgsSpatiaLiteUtils.
