+++ src/providers/spatialite/qgsspatialiteutils.cpp
@@ -23,6 +23,7 @@
 #include "qgslogger.h"
 #include "qgsmessagelog.h"
 #include "qgsproject.h"
+#include <qgsjsonutils.h>
 
 //-----------------------------------------------------------
 //  SpatialiteDbInfo
@@ -6458,6 +6459,7 @@ QgsWkbTypes::Type SpatialiteDbLayer::GetGeometryTypeLegacy( const QString spatia
   }
   return geometryType;
 }
+//-----------------------------------------------------------
 QgsWkbTypes::Type SpatialiteDbLayer::GetGeometryType( const int spatialiteGeometryType,   const int spatialiteGeometryDimension )
 {
   Q_UNUSED( spatialiteGeometryDimension );
@@ -6465,7 +6467,618 @@ QgsWkbTypes::Type SpatialiteDbLayer::GetGeometryType( const int spatialiteGeomet
   return geometryType;
 }
 //-----------------------------------------------------------
+// QgsSpatiaLiteProvider functions
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::addLayerFeatures( QgsFeatureList &flist, QgsFeatureSink::Flags flags, QString &errorMessage )
+{
+  int ret = SQLITE_ABORT; // Container not supported
+  if ( isLayerSpatialite() )
+  {
+    sqlite3_stmt *stmt = nullptr;
+    char *errMsg = nullptr;
+    bool toCommit = false;
+    QString sql;
+    QString values;
+    QString separator;
+    int ia;
+    if ( flist.isEmpty() )
+      return true;
+    QgsAttributes attributevec = flist[0].attributes();
+
+    ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+    if ( ret == SQLITE_OK )
+    {
+      toCommit = true;
+
+      sql = QStringLiteral( "INSERT INTO %1(" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( getTableName() ) );
+      values = QStringLiteral( ") VALUES (" );
+      separator = QLatin1String( "" );
+
+      if ( !getGeometryColumn().isEmpty() )
+      {
+        sql += separator + QgsSpatiaLiteUtils::quotedIdentifier( getGeometryColumn() );
+        values += separator + geomParam();
+        separator = ',';
+      }
+
+      for ( int i = 0; i < attributevec.count(); ++i )
+      {
+        if ( i >= getAttributeFields().count() )
+          continue;
+
+        QString fieldname = getAttributeFields().at( i ).name();
+        // do not write the Primary Key, will be set with auto increment [would overwrite a previously set record.]
+        if ( ( fieldname.isEmpty() ) || ( fieldname == getGeometryColumn() ) || ( fieldname == getPrimaryKey() ) )
+          continue;
+
+        sql += separator + QgsSpatiaLiteUtils::quotedIdentifier( fieldname );
+        values += separator + '?';
+        separator = ',';
+      }
+      sql += values;
+      sql += ')';
+      // SQLite prepared statement
+      ret = sqlite3_prepare_v2( dbSqliteHandle(), sql.toUtf8().constData(), -1, &stmt, nullptr );
+      if ( ret == SQLITE_OK )
+      {
+        for ( QgsFeatureList::iterator feature = flist.begin(); feature != flist.end(); ++feature )
+        {
+          // looping on each feature to insert
+          QgsAttributes attributevec = feature->attributes();
+
+          // resetting Prepared Statement and bindings
+          sqlite3_reset( stmt );
+          sqlite3_clear_bindings( stmt );
+
+          // initializing the column counter
+          ia = 0;
+
+          if ( !getGeometryColumn().isEmpty() )
+          {
+            // binding GEOMETRY to Prepared Statement
+            if ( !feature->hasGeometry() )
+            {
+              sqlite3_bind_null( stmt, ++ia );
+            }
+            else
+            {
+              unsigned char *wkb = nullptr;
+              int wkb_size;
+              QByteArray featureWkb = feature->geometry().exportToWkb();
+              // convertFromGeosWKB( reinterpret_cast<const unsigned char *>( featureWkb.constData() ),featureWkb.length(),&wkb, &wkb_size, nDims );
+              QgsSpatiaLiteUtils::convertFromGeosWKB( reinterpret_cast<const unsigned char *>( featureWkb.constData() ), featureWkb.length(), &wkb, &wkb_size, getCoordDimensions() );
+              if ( !wkb )
+                sqlite3_bind_null( stmt, ++ia );
+              else
+                sqlite3_bind_blob( stmt, ++ia, wkb, wkb_size, QgsSpatiaLiteUtils::deleteWkbBlob );
+            }
+          }
+
+          for ( int i = 0; i < attributevec.count(); ++i )
+          {
+            QVariant v = attributevec.at( i );
+
+            // binding values for each attribute
+            if ( i >= getAttributeFields().count() )
+              break;
+
+            QString fieldname = getAttributeFields().at( i ).name();
+            // do not write the Primary Key, will be set with auto increment [would overwrite a previously set record.]
+            if ( ( fieldname.isEmpty() ) || ( fieldname == getGeometryColumn() ) || ( fieldname == getPrimaryKey() ) )
+              continue;
+
+            QVariant::Type type = getAttributeFields().at( i ).type();
+            if ( !v.isValid() )
+            {
+              ++ia;
+            }
+            else if ( v.isNull() )
+            {
+              // binding a NULL value
+              sqlite3_bind_null( stmt, ++ia );
+            }
+            else if ( type == QVariant::Int )
+            {
+              // binding an INTEGER value
+              sqlite3_bind_int( stmt, ++ia, v.toInt() );
+            }
+            else if ( type == QVariant::LongLong )
+            {
+              // binding a LONGLONG value
+              sqlite3_bind_int64( stmt, ++ia, v.toLongLong() );
+            }
+            else if ( type == QVariant::Double )
+            {
+              // binding a DOUBLE value
+              sqlite3_bind_double( stmt, ++ia, v.toDouble() );
+            }
+            else if ( type == QVariant::String )
+            {
+              QString stringVal = v.toString();
+
+              // binding a TEXT value
+              QByteArray ba = stringVal.toUtf8();
+              sqlite3_bind_text( stmt, ++ia, ba.constData(), ba.size(), SQLITE_TRANSIENT );
+            }
+            else if ( type == QVariant::StringList || type == QVariant::List )
+            {
+              const QByteArray ba = QgsJsonUtils::encodeValue( v ).toUtf8();
+              sqlite3_bind_text( stmt, ++ia, ba.constData(), ba.size(), SQLITE_TRANSIENT );
+            }
+            else if ( type == QVariant::QVariant::Date || type == QVariant::DateTime )
+            {
+              QString stringVal = v.toString();
+              const QByteArray ba = stringVal.toUtf8();
+              sqlite3_bind_text( stmt, ++ia, ba.constData(), ba.size(), SQLITE_TRANSIENT );
+            }
+            else
+            {
+              // Unknown type: bind a NULL value
+              sqlite3_bind_null( stmt, ++ia );
+            }
+          }
+          // performing actual row insert
+          ret = sqlite3_step( stmt );
+          if ( ret == SQLITE_DONE || ret == SQLITE_ROW )
+          {
+            // update feature id
+            if ( !( flags & QgsFeatureSink::FastInsert ) )
+            {
+              feature->setId( sqlite3_last_insert_rowid( dbSqliteHandle() ) );
+            }
+            //mNumberFeatures++;// Done in  getLayerExtent
+          }
+          else
+          {
+            // some unexpected error occurred
+            const char *err = sqlite3_errmsg( dbSqliteHandle() );
+            errMsg = ( char * ) sqlite3_malloc( ( int ) strlen( err ) + 1 );
+            strcpy( errMsg, err );
+            break;
+          }
+        }
+        sqlite3_finalize( stmt );
+        if ( ret == SQLITE_DONE || ret == SQLITE_ROW )
+        {
+          ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+        }
+      } // prepared statement
+    } // BEGIN statement
+    if ( ret != SQLITE_OK )
+    {
+      errorMessage = QString( tr( "SpatialiteDbLayer::addLayerFeatures: SQLite error: %2 rc=%3\nSQL: %1" ).arg( sql, errMsg ? errMsg : tr( "unknown cause" ) ).arg( ret ) );
+      if ( errMsg )
+      {
+        sqlite3_free( errMsg );
+      }
+      if ( toCommit )
+      {
+        // ROLLBACK after some previous error
+        ( void )sqlite3_exec( dbSqliteHandle(), "ROLLBACK", nullptr, nullptr, nullptr );
+      }
+    }
+    getLayerExtent( true, true );
+  }
+  return ret == SQLITE_OK;
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::deleteLayerFeatures( const QgsFeatureIds &id, QString &errorMessage )
+{
+  sqlite3_stmt *stmt = nullptr;
+  char *errMsg = nullptr;
+  QString sql;
+
+  int ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg );
+    return false;
+  }
+
+  sql = QStringLiteral( "DELETE FROM %1 WHERE %2=?" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( getTableName() ), QgsSpatiaLiteUtils::quotedIdentifier( getPrimaryKey() ) );
+
+  // SQLite prepared statement
+  if ( sqlite3_prepare_v2( dbSqliteHandle(), sql.toUtf8().constData(), -1, &stmt, nullptr ) != SQLITE_OK )
+  {
+    // some error occurred
+    errorMessage = QString( tr( "SQLite error: %2\nSQL: %1" ).arg( sql, sqlite3_errmsg( dbSqliteHandle() ) ) );
+    return false;
+  }
+
+  for ( QgsFeatureIds::const_iterator it = id.begin(); it != id.end(); ++it )
+  {
+    // looping on each feature to be deleted
+    // resetting Prepared Statement and bindings
+    sqlite3_reset( stmt );
+    sqlite3_clear_bindings( stmt );
+
+    qint64 fid = FID_TO_NUMBER( *it );
+    sqlite3_bind_int64( stmt, 1, fid );
+
+    // performing actual row deletion
+    ret = sqlite3_step( stmt );
+    if ( ret == SQLITE_DONE || ret == SQLITE_ROW )
+    {
+      // mNumberFeatures--; will be set with getLayerExtent(true, true );
+    }
+    else
+    {
+      // some unexpected error occurred
+      const char *err = sqlite3_errmsg( dbSqliteHandle() );
+      errMsg = ( char * ) sqlite3_malloc( ( int ) strlen( err ) + 1 );
+      strcpy( errMsg, err );
+      handleError( sql, errMsg, true );
+      return false;
+    }
+  }
+  sqlite3_finalize( stmt );
+
+  ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg, true );
+    return false;
+  }
+  getLayerExtent( true, true );
+  return true;
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::truncateLayerTableRows()
+{
+  char *errMsg = nullptr;
+  QString sql;
+
+  int ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg );
+    return false;
+  }
+
+  sql = QStringLiteral( "DELETE FROM %1" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( getTableName() ) );
+  ret = sqlite3_exec( dbSqliteHandle(), sql.toUtf8().constData(), nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg, true );
+    return false;
+  }
+
+  ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg, true );
+    return false;
+  }
+  getLayerExtent( true, true );
+  return true;
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::changeLayerGeometryValues( const QgsGeometryMap &geometry_map )
+{
+  int ret = SQLITE_ABORT; // Container not supported
+  if ( isLayerSpatialite() )
+  {
+    sqlite3_stmt *stmt = nullptr;
+    char *errMsg = nullptr;
+    QString sql;
+
+    ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg );
+      return false;
+    }
+    sql =
+      QStringLiteral( "UPDATE %1 SET %2=GeomFromWKB(?, %3) WHERE %4=?" )
+      .arg( QgsSpatiaLiteUtils::quotedIdentifier( getTableName() ),
+            QgsSpatiaLiteUtils::quotedIdentifier( getGeometryColumn() ) )
+      .arg( getSrid() )
+      .arg( QgsSpatiaLiteUtils::quotedIdentifier( getPrimaryKey() ) );
+
+    // SQLite prepared statement
+    if ( sqlite3_prepare_v2( dbSqliteHandle(), sql.toUtf8().constData(), -1, &stmt, nullptr ) != SQLITE_OK )
+    {
+      // some error occurred
+      QgsMessageLog::logMessage( tr( "SQLite error: %2\nSQL: %1" ).arg( sql, sqlite3_errmsg( dbSqliteHandle() ) ), tr( "SpatiaLite" ) );
+      return false;
+    }
+
+    for ( QgsGeometryMap::const_iterator iter = geometry_map.constBegin(); iter != geometry_map.constEnd(); ++iter )
+    {
+      // resetting Prepared Statement and bindings
+      sqlite3_reset( stmt );
+      sqlite3_clear_bindings( stmt );
+
+      // binding GEOMETRY to Prepared Statement
+      unsigned char *wkb = nullptr;
+      int wkb_size;
+      QByteArray iterWkb = iter->exportToWkb();
+      QgsSpatiaLiteUtils::convertFromGeosWKB( reinterpret_cast<const unsigned char *>( iterWkb.constData() ), iterWkb.length(), &wkb, &wkb_size, getCoordDimensions() );
+      if ( !wkb )
+        sqlite3_bind_null( stmt, 1 );
+      else
+        sqlite3_bind_blob( stmt, 1, wkb, wkb_size, QgsSpatiaLiteUtils::deleteWkbBlob );
+      sqlite3_bind_int64( stmt, 2, FID_TO_NUMBER( iter.key() ) );
+
+      // performing actual row update
+      ret = sqlite3_step( stmt );
+      if ( ret == SQLITE_DONE || ret == SQLITE_ROW )
+        ;
+      else
+      {
+        // some unexpected error occurred
+        const char *err = sqlite3_errmsg( dbSqliteHandle() );
+        errMsg = ( char * ) sqlite3_malloc( ( int ) strlen( err ) + 1 );
+        strcpy( errMsg, err );
+        handleError( sql, errMsg, true );
+        return false;
+      }
+    }
+    sqlite3_finalize( stmt );
+    ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg, true );
+      return false;
+    }
+  }
+  return ret == SQLITE_OK;
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::addLayerAttributes( const QList<QgsField> &attributes )
+{
+  char *errMsg = nullptr;
+  QString sql;
+
+  if ( attributes.isEmpty() )
+    return true;
+
+  int ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg );
+    return false;
+  }
+
+  for ( QList<QgsField>::const_iterator iter = attributes.begin(); iter != attributes.end(); ++iter )
+  {
+    sql = QStringLiteral( "ALTER TABLE \"%1\" ADD COLUMN \"%2\" %3" )
+          .arg( getTableName(),
+                iter->name(),
+                iter->typeName() );
+    ret = sqlite3_exec( dbSqliteHandle(), sql.toUtf8().constData(), nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg, true );
+      return false;
+    }
+  }
+
+  ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg, true );
+    return false;
+  }
+  getNumberFeatures( true );
+  // reload columns
+  getCapabilities( true );
+
+  return true;
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::changeLayerAttributeValues( const QgsChangedAttributesMap &attr_map )
+{
+  char *errMsg = nullptr;
+  QString sql;
+
+  if ( attr_map.isEmpty() )
+    return true;
+
+  int ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg );
+    return false;
+  }
+
+  for ( QgsChangedAttributesMap::const_iterator iter = attr_map.begin(); iter != attr_map.end(); ++iter )
+  {
+    // Loop over all changed features
+
+    QgsFeatureId fid = iter.key();
+
+    // skip added features
+    if ( FID_IS_NEW( fid ) )
+      continue;
+
+    const QgsAttributeMap &attrs = iter.value();
+    if ( attrs.isEmpty() )
+      continue;
+
+    QString sql = QStringLiteral( "UPDATE %1 SET " ).arg( QgsSpatiaLiteUtils::quotedIdentifier( getTableName() ) );
+    bool first = true;
+
+    // cycle through the changed attributes of the feature
+    for ( QgsAttributeMap::const_iterator siter = attrs.begin(); siter != attrs.end(); ++siter )
+    {
+      // Loop over all changed attributes
+      try
+      {
+        QgsField fld = getAttributeField( siter.key() );
+        const QVariant &val = siter.value();
+
+        if ( !first )
+          sql += ',';
+        else
+          first = false;
+
+        QVariant::Type type = fld.type();
+
+        if ( val.isNull() || !val.isValid() )
+        {
+          // binding a NULL value
+          sql += QStringLiteral( "%1=NULL" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( fld.name() ) );
+        }
+        else if ( type == QVariant::Int || type == QVariant::LongLong || type == QVariant::Double )
+        {
+          // binding a NUMERIC value
+          sql += QStringLiteral( "%1=%2" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( fld.name() ), val.toString() );
+        }
+        else if ( type == QVariant::StringList || type == QVariant::List )
+        {
+          // binding an array value
+          sql += QStringLiteral( "%1=%2" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( fld.name() ), QgsSpatiaLiteUtils::quotedValue( QgsJsonUtils::encodeValue( val ) ) );
+        }
+        else
+        {
+          // binding a TEXT value
+          sql += QStringLiteral( "%1=%2" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( fld.name() ), QgsSpatiaLiteUtils::quotedValue( val.toString() ) );
+        }
+      }
+      catch ( QgsSpatiaLiteUtils::SLFieldNotFound )
+      {
+        // Field was missing - shouldn't happen
+      }
+    }
+    sql += QStringLiteral( " WHERE %1=%2" ).arg( QgsSpatiaLiteUtils::quotedIdentifier( getPrimaryKey() ) ).arg( fid );
+
+    ret = sqlite3_exec( dbSqliteHandle(), sql.toUtf8().constData(), nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg, true );
+      return false;
+    }
+  }
+
+  ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+  if ( ret != SQLITE_OK )
+  {
+    handleError( sql, errMsg, true );
+    return false;
+  }
+
+  return true;
+}
+//-----------------------------------------------------------
+QgsField SpatialiteDbLayer::getAttributeField( int index ) const
+{
+  if ( index < 0 || index >= getAttributeFields().count() )
+  {
+    QgsMessageLog::logMessage( tr( "FAILURE: Field %1 not found." ).arg( index ), tr( "SpatiaLite" ) );
+    throw QgsSpatiaLiteUtils::SLFieldNotFound();
+  }
+
+  return getAttributeFields().at( index );
+}
+//-----------------------------------------------------------
+bool SpatialiteDbLayer::createLayerAttributeIndex( int field )
+{
+  int ret = SQLITE_ABORT; // Container not supported
+  if ( isLayerSpatialite() )
+  {
+    char *errMsg = nullptr;
+
+    if ( field < 0 || field >= getAttributeFields().count() )
+      return false;
+
+    QString sql;
+    QString fieldName;
+
+    ret = sqlite3_exec( dbSqliteHandle(), "BEGIN", nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg );
+      return false;
+    }
+
+    fieldName = getAttributeFields().at( field ).name();
+
+    sql = QStringLiteral( "CREATE INDEX IF NOT EXISTS %1 ON \"%2\" (%3)" )
+          .arg( QgsSpatiaLiteUtils::createIndexName( getTableName(), fieldName ),
+                getTableName(),
+                QgsSpatiaLiteUtils::quotedIdentifier( fieldName ) );
+    ret = sqlite3_exec( dbSqliteHandle(), sql.toUtf8().constData(), nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg, true );
+      return false;
+    }
+
+    ret = sqlite3_exec( dbSqliteHandle(), "COMMIT", nullptr, nullptr, &errMsg );
+    if ( ret != SQLITE_OK )
+    {
+      handleError( sql, errMsg, true );
+      return false;
+    }
+  }
+  return ret == SQLITE_OK;
+}
+//-----------------------------------------------------------
+QString SpatialiteDbLayer::geomParam() const
+{
+  QString geometry;
+  bool forceMulti = QgsWkbTypes::isMultiType( getGeometryType() );
+  // ST_Multi function is available from QGIS >= 2.4
+  bool hasMultiFunction = dbSpatialiteVersionMajor() > 2 ||
+                          ( dbSpatialiteVersionMajor() == 2 && dbSpatialiteVersionMinor() >= 4 );
+
+  if ( forceMulti && hasMultiFunction )
+  {
+    geometry += QLatin1String( "ST_Multi(" );
+  }
+  geometry += QStringLiteral( "GeomFromWKB(?, %2)" ).arg( mSrid );
+  if ( forceMulti && hasMultiFunction )
+  {
+    geometry += ')';
+  }
+  return geometry;
+}
+//-----------------------------------------------------------
+void SpatialiteDbLayer::handleError( const QString &sql, char *errorMessage, bool rollback )
+{
+  QgsMessageLog::logMessage( tr( "SQLite error: %2\nSQL: %1" ).arg( sql, errorMessage ? errorMessage : tr( "unknown cause" ) ), tr( "SpatiaLite" ) );
+  // unexpected error
+  if ( errorMessage )
+  {
+    sqlite3_free( errorMessage );
+  }
+
+  if ( rollback )
+  {
+    // ROLLBACK after some previous error
+    ( void )sqlite3_exec( dbSqliteHandle(), "ROLLBACK", nullptr, nullptr, nullptr );
+  }
+}
+//-----------------------------------------------------------
 //  QgsSpatiaLiteUtils functions
+//-----------------------------------------------------------
+QString QgsSpatiaLiteUtils::createIndexName( QString tableName, QString field )
+{
+  QRegularExpression safeExp( QStringLiteral( "[^a-zA-Z0-9]" ) );
+  tableName.replace( safeExp, QStringLiteral( "_" ) );
+  field.replace( safeExp, QStringLiteral( "_" ) );
+  return QStringLiteral( "%1_%2_idx" ).arg( tableName, field );
+}
+QString QgsSpatiaLiteUtils::quotedIdentifier( QString id )
+{
+  id.replace( '\"', QLatin1String( "\"\"" ) );
+  return id.prepend( '\"' ).append( '\"' );
+}
+//-----------------------------------------------------------
+QString QgsSpatiaLiteUtils::quotedValue( QString value )
+{
+  if ( value.isNull() )
+    return QStringLiteral( "NULL" );
+
+  value.replace( '\'', QLatin1String( "''" ) );
+  return value.prepend( '\'' ).append( '\'' );
+}
+//-----------------------------------------------------------
+void QgsSpatiaLiteUtils::deleteWkbBlob( void *wkbBlob )
+{
+  delete[]( char * )wkbBlob;
+}
 //-----------------------------------------------------------
 SpatialiteDbInfo *QgsSpatiaLiteUtils::GetSpatialiteDbInfoWrapper( QString sDatabaseFileName, QString sLayerName, bool bLoadLayers, sqlite3 *sqlite_handle )
 {
